suppress warnings during build
PPCi64 builds of dropwatch currently fail with the latest gcc as they warn about
type mismatches in a few prinfs between __u64 and unsigned long long types.  Not
sure why x86[_64] doesn't gripe here, but the conversion is fine, so lets do the
cast and quiet the compiler

+++ src/main.c
@@ -333,7 +333,7 @@ void handle_dm_alert_msg(struct netlink_message *msg, int err)
 			printf ("%d drops at location %p\n", alert->points[i].count, location);
 		else
 			printf ("%d drops at %s+%llx (%p)\n",
-				alert->points[i].count, res.symbol, res.offset, location);
+				alert->points[i].count, res.symbol, (unsigned long long)res.offset, location);
 		acount++;
 		if (alimit && (acount == alimit)) {
 			printf("Alert limit reached, deactivating!\n");
